LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-basic-expected.txt


PASS --pf
PASS --pf flip-block
PASS --pf flip-inline
PASS --pf flip-block flip-inline
PASS --pf flip-inline flip-block
FAIL --pf flip-start assert_equals: offsetLeft expected 20 but got 10
FAIL --pf flip-block flip-start flip-inline assert_equals: offsetLeft expected 20 but got 360
FAIL --pf flip-inline flip-start flip-block assert_equals: offsetLeft expected 20 but got 360
FAIL --pf flip-start flip-block assert_equals: offsetLeft expected 20 but got 10
FAIL --pf flip-inline flip-start assert_equals: offsetLeft expected 20 but got 360
FAIL --pf flip-start flip-inline assert_equals: offsetLeft expected 340 but got 360
FAIL --pf flip-block flip-start assert_equals: offsetLeft expected 340 but got 10
FAIL --pf flip-start flip-block flip-inline assert_equals: offsetLeft expected 340 but got 360
FAIL --pf flip-start flip-inline flip-block assert_equals: offsetLeft expected 340 but got 360
FAIL --pf flip-inline flip-block flip-start assert_equals: offsetLeft expected 340 but got 360
FAIL --pf flip-block flip-inline flip-start assert_equals: offsetLeft expected 340 but got 360
